+++ src/main/java/christmas/view/validation/MenuValidator.java
@@ -7,7 +7,7 @@ import static christmas.view.constant.InputConstant.ORDER_DELIMITER;
 
 import christmas.model.domain.constant.Menu;
 import christmas.model.domain.constant.Type;
-import christmas.view.Mapper;
+import christmas.view.mapper.Mapper;
 import christmas.view.constant.ErrorMessage;
 import java.util.Arrays;
 import java.util.EnumMap;
